Adds router
