+++ package.json
@@ -12,7 +12,9 @@
     "axios": "^0.25.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-is": "^17.0.2",
     "react-modal": "^3.14.4",
+    "react-router-dom": "^6.2.1",
     "styled-components": "^5.3.3"
   },
   "devDependencies": {
